fix grad again

+++ guided_diffusion/respace_jittor.py
@@ -121,7 +121,7 @@ class _WrappedModel:
         self.original_num_steps = original_num_steps
 
     def __call__(self, x, ts, **kwargs):
-        map_tensor = jt.Var(self.timestep_map, dtype=ts.dtype)
+        map_tensor = jt.Var(self.timestep_map)
         new_ts = map_tensor[ts]
         if self.rescale_timesteps:
             new_ts = new_ts.float() * (1000.0 / self.original_num_steps)
